Refactor service and mapper return types to Integer
Changed service and mapper methods to use Integer instead of Long for insert, update, delete, and count operations. Added BuildServiceImpl and ProductInfoServiceImpl for service implementation generation. Updated method naming conventions for clarity and consistency. Adjusted related templates and entity classes to use Integer types for pagination and count fields. Removed unused XML mapper files.

+++ sql2codeInitDemo/src/main/java/com/slilio/sql2codeInitDemo/entity/vo/PaginationResultVO.java
@@ -17,8 +17,7 @@ public class PaginationResultVO<T> {
     this.list = list;
   }
 
-  public PaginationResultVO(
-      Integer totalCount, Integer pageSize, Integer pageNo, Integer pageTotal, List<T> list) {
+  public PaginationResultVO(Integer totalCount, Integer pageSize, Integer pageNo, Integer pageTotal, List<T> list) {
     if (pageNo == 0) {
       pageNo = 1;
     }
@@ -35,11 +34,11 @@ public class PaginationResultVO<T> {
 
   public PaginationResultVO() {}
 
-  public Integer getTotalCount() {
+  public Long getTotalCount() {
     return totalCount;
   }
 
-  public void setTotalCount(Integer totalCount) {
+  public void setTotalCount(Long totalCount) {
     this.totalCount = totalCount;
   }
